Trace Guard/requirements.txt
flask==2.3.2
pandas==2.0.3
tweepy==4.14.0
instaloader==4.9.3
telethon==1.28.5
nest-asyncio==1.5.6
transformers==4.31.0
scikit-learn==1.3.0
wordcloud==1.9.2
fpdf2==2.7.5
python-dotenv==1.0.0    